move QSym material back
The proof of the shuffle-compatibility of Epk does not
rely on QSym and on the concept of descent statistics.
Thus, I've moved the definitions of QSym and of descent
statistics to a later section, so that the former
proof is more accessible.

Also, added a few sentences about the dimensions of
the shuffle algebra of Epk.