new section for lacunar sets
This section collects all lemmas about lacunar sets, and proves
them in more details than before.
This also prepares a total order for a different proof of linear
independency that may be added later.
Also, there is now the notation \mathbf{L}_n for the indexing
set of our family.